fix test_dir_rm, test_sync_conflict_path, test_hash_mess and test_cs_folder_conflicts_del

+++ cloudsync/sync/manager.py
@@ -1273,7 +1273,7 @@ class SyncManager(Runnable):
 
                 return REQUEUE  # we don't want to punt here, we just manually adjusted the priority above
 
-        if sync[changed].exists == TRASHED:
+        if sync[changed].exists == TRASHED and sync[synced].exists == EXISTS:
             log.debug("delete")
             return self.delete_synced(sync, changed, synced)
